+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Release NEXT
 ------------
 - Render actions metadata for new resources.
 - Historical resource calculation is made optional.
+- Fix Sentry integration.
 
 Release 0.103.0
 ---------------